tests: move IColSelect and IColsSelect from gristUtils to gristWebDriverUtils
Part of refactor to sync with grist-widget.
